Moved all custom acquisitions to commissioning file.
